+++ tests/fixtures/rules/framework.react.class-component/true_positive/src/Widget.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export class Widget extends React.Component {
+  render() {
+    return <div>hello</div>;
+  }
+}